refactor: audienceviewpage hook분리
