+++ src/pages/AudienceView/AudienceViewPage.jsx
@@ -1,11 +1,5 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import React, { useMemo, useRef, useState } from "react";
+import { useParams } from "react-router-dom";
 import AudiencePanel from "../../components/Audience/AudiencePanel";
 import SidebarSlides from "../../components/SidebarSlides";
 import {
@@ -15,81 +9,31 @@ import {
 } from "./AudienceViewPage.styles";
 import SlideViewer from "../../components/Audience/SlideViewer_audience/SlideViewer_audience";
 import EmojiPanel from "../../components/Audience/EmojiPanel";
-import { joinRoom } from "../../services/roomService";
-import websocketService, {
-  WebSocketService,
-} from "../../services/websocketService";
-import { fetchAllOriginalSlideUrls } from "../../services/presentationService";
-import { getStickersByAudience } from "../../services/stickerService";
+import { WebSocketService } from "../../services/websocketService";
 import useAudienceQuestions from "../../hooks/useAudienceQuestions";
 import useEmojiReactions from "../../hooks/useEmojiReactions";
 import useStickerLoader from "../../hooks/useStickerLoader";
 import useAudienceJoinRoom from "../../hooks/useAudienceJoinRoom";
 import usePresenterPageSync from "../../hooks/usePresenterPageSync";
-import SELECTED_EMOJI_ICONS from "../../constants/emojiIcons";
+import useAudienceSlides from "../../hooks/useAudienceSlides";
+import useAudienceSlideNavigation from "../../hooks/useAudienceSlideNavigation";
+import useAudienceWebSocketSubscriptions from "../../hooks/useAudienceWebSocketSubscriptions";
+import useAudienceEventHandlers from "../../hooks/useAudienceEventHandlers";
+import useAudienceInitialState from "../../hooks/useAudienceInitialState";
+import useAudienceFocusHighlight from "../../hooks/useAudienceFocusHighlight";
 import DelayAudience from "./DelayAudience";
 
 const AudienceViewPage = () => {
   const { code } = useParams();
-  const navigate = useNavigate();
 
-  // 세션 스토리지에서 초기 sessionStatus 확인
-  const getInitialSessionStatus = () => {
-    if (!code) return "waiting";
-    try {
-      const storageKey = `boini_audience_${code}`;
-      const stored = sessionStorage.getItem(storageKey);
-      if (stored) {
-        const storedData = JSON.parse(stored);
-        // audienceId와 audienceToken이 있고 sessionStatus가 있으면 복원
-        if (storedData.audienceId && storedData.audienceToken && storedData.sessionStatus) {
-          return storedData.sessionStatus;
-        }
-      }
-    } catch (_error) {
-      // 세션 스토리지 읽기 실패 시 기본값 사용
-    }
-    return "waiting";
-  };
-
-  // 세션 스토리지에서 초기 슬라이드 위치 확인 (발표자 현재 페이지)
-  const getInitialSlide = () => {
-    if (!code) return 0;
-    try {
-      const storageKey = `boini_audience_${code}`;
-      const stored = sessionStorage.getItem(storageKey);
-      if (stored) {
-        const storedData = JSON.parse(stored);
-        // currentPage가 있으면 인덱스로 변환 (페이지는 1부터 시작, 인덱스는 0부터)
-        if (storedData.currentPage) {
-          const presenterPage = Number(storedData.currentPage);
-          if (Number.isFinite(presenterPage) && presenterPage > 0) {
-            return presenterPage - 1; // 페이지 번호를 인덱스로 변환
-          }
-        }
-      }
-    } catch (_error) {
-      // 세션 스토리지 읽기 실패 시 기본값 사용
-    }
-    return 0;
-  };
-
-  const [slides, setSlides] = useState([]);
-  const [currentSlide, setCurrentSlide] = useState(getInitialSlide);
-  const [followPresenter, setFollowPresenter] = useState(true);
-  const [selectedEmoji, setSelectedEmoji] = useState(null);
-  const [showStamps, setShowStamps] = useState(true);
   const [roomId, setRoomId] = useState(null);
   const [audienceId, setAudienceId] = useState(null);
   const [audienceToken, setAudienceToken] = useState(null);
   const [wsUrl, setWsUrl] = useState(null);
   const [deckId, setDeckId] = useState(null);
   const [totalPages, setTotalPages] = useState(0);
-  const [loadingSlides, setLoadingSlides] = useState(false);
-  const [slidesError, setSlidesError] = useState(null);
-  const [isWebsocketReady, setIsWebsocketReady] = useState(false);
-  const [showFocusHighlight, setShowFocusHighlight] = useState(false);
-  const [sessionStatus, setSessionStatus] = useState(getInitialSessionStatus);
+  const [selectedEmoji, setSelectedEmoji] = useState(null);
+  const [showStamps, setShowStamps] = useState(true);
 
   // 🔹 빠른 설정 옵션 상태 (발표자로부터 받은 설정)
   const [quickSettings, setQuickSettings] = useState({
@@ -104,18 +48,49 @@ const AudienceViewPage = () => {
     presenterPage: null,
   });
 
-  const questionSubscriptionsRef = useRef([]);
-  const pageChangeUnsubscribeRef = useRef(null);
-  const focusOnUnsubscribeRef = useRef(null);
-  const optionUnsubscribeRef = useRef(null);
-  const unlockUnsubscribeRef = useRef(null);
-  const sessionStateUnsubscribeRef = useRef(null);
-  const followPresenterRef = useRef(followPresenter);
-  const initialSlide = getInitialSlide();
-  const prevSlideRef = useRef(initialSlide);
   const lastPresenterPageRef = useRef(0);
-  const focusHighlightTimeoutRef = useRef(null);
-  const hasNavigatedToRatingRef = useRef(false);
+
+  // 초기 상태 관리
+  const { sessionStatus, setSessionStatus } = useAudienceInitialState({
+    code,
+    roomId,
+  });
+
+  // 슬라이드 관리
+  const {
+    slides,
+    showSlidesPlaceholder,
+    waitingMessage,
+    hasSlidesError,
+    handleRetryFetchSlides,
+  } = useAudienceSlides({
+    roomId,
+    deckId,
+    totalPages,
+  });
+
+  // 슬라이드 네비게이션
+  const {
+    currentSlide,
+    setCurrentSlide,
+    followPresenter,
+    setFollowPresenter,
+    followPresenterRef,
+    prevSlideRef,
+    changeCurrentSlide,
+    handleToggleFollowPresenter,
+    handleAudienceSelectSlide,
+  } = useAudienceSlideNavigation({
+    code,
+    slideCount: slides.length,
+    roomId,
+    audienceId,
+    lastPresenterPageRef,
+  });
+
+  // Focus Highlight 관리
+  const { showFocusHighlight, triggerFocusHighlight } =
+    useAudienceFocusHighlight();
 
   const {
     questions,
@@ -145,107 +120,6 @@ const AudienceViewPage = () => {
     service: reactionService,
   });
 
-  const slideCount = slides.length;
-
-  const changeCurrentSlide = useCallback(
-    (
-      nextIndex,
-      {
-        source = "audience",
-        broadcast = true,
-        preserveFollowState = false,
-      } = {}
-    ) => {
-      setCurrentSlide((prev) => {
-        if (!Number.isFinite(nextIndex)) {
-          return prev;
-        }
-
-        const maxIndex = Math.max(slideCount - 1, 0);
-        const clamped = Math.min(Math.max(nextIndex, 0), maxIndex);
-
-        if (clamped === prev) {
-          return prev;
-        }
-
-        // 모든 청중 페이지 변경을 백엔드로 전송 (동기화 포함)
-        if (
-          roomId &&
-          audienceId &&
-          websocketService.getIsConnected()
-        ) {
-          const beforePage = prev + 1;
-          const changedPage = clamped + 1;
-
-          websocketService.sendAudiencePageChange(
-            roomId,
-            audienceId,
-            beforePage,
-            changedPage
-          );
-        }
-
-        prevSlideRef.current = clamped;
-        return clamped;
-      });
-
-      if (!preserveFollowState && source !== "presenter") {
-        setFollowPresenter(false);
-      }
-    },
-    [slideCount, roomId, audienceId]
-  );
-
-  const loadSlides = useCallback(
-    async ({ signal } = {}) => {
-      if (!roomId || !deckId || !totalPages) return;
-      if (signal?.aborted) return;
-
-      setLoadingSlides(true);
-      setSlidesError(null);
-
-      try {
-        const urls = await fetchAllOriginalSlideUrls(
-          roomId,
-          deckId,
-          totalPages
-        );
-
-        if (signal?.aborted) return;
-        setSlides(urls);
-      } catch (error) {
-        if (signal?.aborted) return;
-        setSlidesError(error);
-      } finally {
-        if (signal?.aborted) return;
-        setLoadingSlides(false);
-      }
-    },
-    [roomId, deckId, totalPages]
-  );
-
-  useEffect(() => {
-    followPresenterRef.current = followPresenter;
-  }, [followPresenter]);
-
-  // 🔹 sessionStatus 변경 시 세션 스토리지에 저장
-  useEffect(() => {
-    if (!code || !roomId) return;
-    
-    try {
-      const storageKey = `boini_audience_${code}`;
-      const stored = sessionStorage.getItem(storageKey);
-      if (stored) {
-        const storedData = JSON.parse(stored);
-        // sessionStatus 업데이트
-        storedData.sessionStatus = sessionStatus;
-        sessionStorage.setItem(storageKey, JSON.stringify(storedData));
-      }
-    } catch (_error) {
-      // 세션 스토리지 업데이트 실패 시 무시
-    }
-  }, [sessionStatus, code, roomId]);
-
   // 🔹 방 입장 및 초기 설정 처리
   useAudienceJoinRoom({
     code,
@@ -260,17 +134,6 @@ const AudienceViewPage = () => {
     setUnlockSettings,
     lastPresenterPageRef,
   });
-
-  useEffect(() => {
-    if (!roomId || !deckId || !totalPages) return;
-
-    const controller = new AbortController();
-    loadSlides({ signal: controller.signal });
-
-    return () => {
-      controller.abort();
-    };
-  }, [roomId, deckId, totalPages, loadSlides]);
 
   // 🔹 슬라이드 로드 후 발표자 페이지로 동기화
   usePresenterPageSync({
@@ -292,348 +155,41 @@ const AudienceViewPage = () => {
     prefix: "Audience loadStickers",
   });
 
-  useEffect(() => {
-    if (!roomId || !audienceId || !wsUrl || !audienceToken) {
-      return undefined;
-    }
-
-    setIsWebsocketReady(false);
-
-    const onConnect = () => {
-      setIsWebsocketReady(true);
-    };
-
-    const onError = () => {
-      setIsWebsocketReady(false);
-    };
-
-    websocketService.connect(wsUrl, audienceToken, onConnect, onError);
-
-    return () => {
-      setIsWebsocketReady(false);
-      websocketService.disconnect();
-    };
-  }, [roomId, audienceId, wsUrl, audienceToken]);
-
-  useEffect(() => {
-    questionSubscriptionsRef.current.forEach((unsubscribe) => {
-      if (typeof unsubscribe === "function") {
-        unsubscribe();
-      }
-    });
-    questionSubscriptionsRef.current = [];
-
-    pageChangeUnsubscribeRef.current?.();
-    pageChangeUnsubscribeRef.current = null;
-
-    if (!isWebsocketReady || !roomId) {
-      return undefined;
-    }
-
-    const topics = Array.isArray(questionTopics) ? questionTopics : [];
-
-    topics.forEach((topic) => {
-      const unsubscribe = websocketService.subscribe(
-        topic,
-        handleIncomingQuestion
-      );
-      if (typeof unsubscribe === "function") {
-        questionSubscriptionsRef.current.push(unsubscribe);
-      }
-    });
-
-    const pageChangeTopic = `/topic/presentation/${roomId}/pageChange`;
-    pageChangeUnsubscribeRef.current = websocketService.subscribe(
-      pageChangeTopic,
-      (data) => {
-        if (!data || data.changedPage === undefined) {
-          return;
-        }
-
-        const newSlideIndex = Number(data.changedPage);
-        if (Number.isFinite(newSlideIndex)) {
-          lastPresenterPageRef.current = newSlideIndex;
-        }
-
-        if (!followPresenterRef.current) {
-          return;
-        }
-
-        changeCurrentSlide(newSlideIndex, {
-          source: "presenter",
-          broadcast: false,
-          preserveFollowState: true,
-        });
-      }
-    );
-
-    // 🔹 옵션 변경 구독 (리액션 스티커, 질문, 실시간 피드백)
-    const optionTopic = `/topic/presentation/${roomId}/option`;
-    optionUnsubscribeRef.current = websocketService.subscribe(
-      optionTopic,
-      (data) => {
-        if (data) {
-          const payload = data?.data ?? data;
-          const normalized = {
-            sticker: String(payload?.sticker) === "true",
-            question: String(payload?.question) === "true",
-            feedback: String(payload?.feedback) === "true",
-          };
-
-          setQuickSettings(normalized);
-        }
-      }
-    );
-
-    // 🔹 다음 슬라이드 공개 옵션 구독
-    const unlockTopic = `/topic/presentation/${roomId}/option/unlock`;
-    unlockUnsubscribeRef.current = websocketService.subscribe(
-      unlockTopic,
-      (data) => {
-        if (data) {
-          const payload = data?.data ?? data;
-          const maxPage = Number(payload?.maxRevealedPage);
-          const revealAll = String(payload?.revealAllSlides) === "true";
-          const presenterPage = Number(payload?.presenterPage);
-
-          setUnlockSettings({
-            maxRevealedPage: maxPage,
-            revealAllSlides: revealAll,
-            totalPages: Number(payload?.totalPages),
-            presenterPage,
-          });
-        }
-      }
-    );
-
-    // 🔹 세션 상태 변경 구독 (세션 종료 포함)
-    const sessionStateTopic = `/topic/p/${roomId}/public`;
-    sessionStateUnsubscribeRef.current = websocketService.subscribe(
-      sessionStateTopic,
-      (data) => {
-        console.log("[AudienceView] 세션 상태 변경 수신:", data);
-        if (data && data.type === "SESSION_STATE" && data.status) {
-          setSessionStatus(data.status);
-          // 세션 종료 상태 확인
-          if (data.status === "ended" || data.status === "ENDED") {
-            console.log(
-              "[AudienceView] 세션 종료 상태 확인 - rating 페이지로 이동"
-            );
-            if (!hasNavigatedToRatingRef.current) {
-              hasNavigatedToRatingRef.current = true;
-              const targetUrl = code
-                ? `/rating/${encodeURIComponent(code)}`
-                : "/rating";
-              navigate(targetUrl, {
-                replace: false,
-                state: {
-                  roomId: roomId,
-                  audienceId: audienceId,
-                  deckId: deckId,
-                },
-              });
-            }
-          }
-        }
-      }
-    );
-
-    return () => {
-      questionSubscriptionsRef.current.forEach((unsubscribe) => {
-        if (typeof unsubscribe === "function") {
-          unsubscribe();
-        }
-      });
-      questionSubscriptionsRef.current = [];
-
-      pageChangeUnsubscribeRef.current?.();
-      pageChangeUnsubscribeRef.current = null;
-
-      optionUnsubscribeRef.current?.();
-      optionUnsubscribeRef.current = null;
-
-      unlockUnsubscribeRef.current?.();
-      unlockUnsubscribeRef.current = null;
-
-      sessionStateUnsubscribeRef.current?.();
-      sessionStateUnsubscribeRef.current = null;
-    };
-  }, [
-    isWebsocketReady,
+  // WebSocket 구독 관리
+  const { isWebsocketReady } = useAudienceWebSocketSubscriptions({
     roomId,
     audienceId,
-    handleIncomingQuestion,
+    wsUrl,
+    audienceToken,
     questionTopics,
+    handleIncomingQuestion,
     changeCurrentSlide,
     currentSlide,
-  ]);
+    followPresenterRef,
+    lastPresenterPageRef,
+    setSessionStatus,
+    setQuickSettings,
+    setUnlockSettings,
+    code,
+    deckId,
+    triggerFocusHighlight,
+  });
 
-  useEffect(() => {
-    focusOnUnsubscribeRef.current?.();
-    focusOnUnsubscribeRef.current = null;
-
-    if (!isWebsocketReady || !roomId || !websocketService.getIsConnected()) {
-      return undefined;
-    }
-
-    const topic = `/topic/presentation/${roomId}/focusOn`;
-    const unsubscribe = websocketService.subscribeText(topic, (rawMessage) => {
-      if (rawMessage == null) {
-        return;
-      }
-
-      const trimmed = String(rawMessage).trim();
-      if (!trimmed) {
-        return;
-      }
-
-      let parsedNumber;
-      try {
-        parsedNumber = Number(JSON.parse(trimmed));
-        if (!Number.isFinite(parsedNumber)) {
-          throw new Error("NaN");
-        }
-      } catch (error) {
-        parsedNumber = Number(trimmed.replace(/^"+|"+$/g, ""));
-      }
-
-      if (!Number.isFinite(parsedNumber) || parsedNumber < 0) {
-        return;
-      }
-
-      lastPresenterPageRef.current = parsedNumber;
-
-      if (!followPresenterRef.current) {
-        setFollowPresenter(true);
-        followPresenterRef.current = true;
-      }
-
-      setShowFocusHighlight(true);
-      if (focusHighlightTimeoutRef.current) {
-        clearTimeout(focusHighlightTimeoutRef.current);
-      }
-      focusHighlightTimeoutRef.current = setTimeout(() => {
-        setShowFocusHighlight(false);
-        focusHighlightTimeoutRef.current = null;
-      }, 1000);
-
-      changeCurrentSlide(parsedNumber, {
-        source: "focusOn",
-        broadcast: true,
-        preserveFollowState: true,
-      });
+  // 이벤트 핸들러
+  const { handleSelectEmoji, handlePlaceStamp, handleToggleShowStamps } =
+    useAudienceEventHandlers({
+      selectedEmoji,
+      setSelectedEmoji,
+      reactionsReady,
+      roomId,
+      audienceId,
+      wsUrl,
+      currentSlide,
+      addLocalStamp,
+      showStamps,
+      setShowStamps,
     });
 
-    focusOnUnsubscribeRef.current = unsubscribe;
-
-    return () => {
-      if (typeof unsubscribe === "function") {
-        unsubscribe();
-      }
-      focusOnUnsubscribeRef.current = null;
-    };
-  }, [isWebsocketReady, roomId, changeCurrentSlide]);
-
-  // Note: 세션 종료는 /topic/p/{roomId}/public의 SESSION_STATE로 처리됨
-  // 위의 sessionStateTopic 구독에서 status: "ended" 확인하여 rating 페이지로 이동
-
-  useEffect(() => {
-    return () => {
-      if (focusHighlightTimeoutRef.current) {
-        clearTimeout(focusHighlightTimeoutRef.current);
-        focusHighlightTimeoutRef.current = null;
-      }
-    };
-  }, []);
-
-  // 🔹 방향키로 슬라이드 이동
-  useEffect(() => {
-    const handleKeyDown = (event) => {
-      if (event.key === "ArrowLeft" || event.key === "ArrowUp") {
-        // 이전 슬라이드
-        changeCurrentSlide(currentSlide - 1, { source: "audience" });
-      } else if (event.key === "ArrowRight" || event.key === "ArrowDown") {
-        // 다음 슬라이드
-        changeCurrentSlide(currentSlide + 1, { source: "audience" });
-      }
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [currentSlide, changeCurrentSlide]);
-
-  const handleSelectEmoji = (emoji) => setSelectedEmoji(emoji);
-
-  const handlePlaceStamp = ({ xPct, yPct }) => {
-    if (!selectedEmoji || !reactionsReady || !roomId || !audienceId || !wsUrl) {
-      return;
-    }
-
-    if (selectedEmoji.id >= 1 && selectedEmoji.id <= 8) {
-      const now = new Date().toISOString();
-
-      const destination = `/app/presentation/${roomId}/reaction`;
-      const message = {
-        emoji: selectedEmoji.id,
-        audienceID: audienceId,
-        created_at: now,
-        x: xPct,
-        y: yPct,
-        slide: currentSlide + 1,
-      };
-
-      websocketService.send(destination, message);
-
-      const stickerSrc = SELECTED_EMOJI_ICONS[selectedEmoji.id];
-      if (stickerSrc) {
-        addLocalStamp(currentSlide, {
-          id: now,
-          xPct,
-          yPct,
-          src: stickerSrc,
-        });
-      }
-    }
-  };
-
-  const handleToggleFollowPresenter = (checked) => {
-    setFollowPresenter(checked);
-
-    if (checked) {
-      const target = Number.isFinite(lastPresenterPageRef.current)
-        ? lastPresenterPageRef.current
-        : prevSlideRef.current;
-
-      changeCurrentSlide(target, {
-        source: "presenter",
-        broadcast: false,
-        preserveFollowState: true,
-      });
-    }
-  };
-
-  const handleToggleShowStamps = (nextValue) => {
-    setShowStamps(nextValue);
-  };
-
-  const handleAudienceSelectSlide = (slideIndex, options) => {
-    changeCurrentSlide(slideIndex, { source: "audience", ...options });
-  };
-
-  const handleRetryFetchSlides = () => {
-    if (!roomId || !deckId || !totalPages) return;
-    loadSlides();
-  };
-
-  const isSlidesLoading = loadingSlides && slides.length === 0;
-  const hasSlidesError = !loadingSlides && !!slidesError && slides.length === 0;
-  const showSlidesPlaceholder = isSlidesLoading || hasSlidesError;
-  const waitingMessage = hasSlidesError
-    ? "슬라이드를 불러오는 중 오류가 발생했습니다."
-    : "슬라이드를 불러오는 중입니다.";
   const isQuestionListWaiting = questionsLoading && questions.length === 0;
 
   const isSessionWaiting = sessionStatus === "waiting";
